pyproject.toml
[project]
name = "gas-accountant"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "chartengineer>=0.1.2",
    "dfcleaner>=0.1.3",
    "diskcache>=5.6.3",
    "dune-client>=1.7.9",
    "eth-abi>=5.2.0",
    "eth-account>=0.13.7",
    "eth-utils>=5.3.0",
    "flask[async]>=3.1.0",
    "ipykernel>=6.29.5",
    "ipython>=9.2.0",
    "nbformat>=5.10.4",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "prophet>=1.1.6",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "web3>=7.11.0",
]

[tool.setuptools.packages.find]
include = ["gas_accountant", "gas_accountant.*"]
